refactor: page header component
